+++ tests/testthat/test-check.R
@@ -7,3 +7,9 @@ test_that("can determine when to document", {
   ))
   expect_true(can_document(list(roxygennote = packageVersion("roxygen2"))))
 })
+
+test_that("fail instead of sending an email to wrong recipient", {
+  # The testTest package has both Authors@R and Maintainer field - this causes problems in change_maintainer_email().
+  # The function checks if the provided email is actually the one in the maintainer field instead of sending the report to the wrong recipient
+  expect_error(check_win_release(file.path("testTest"), email = "foo@bar.com"))
+})